feat: Support running in Loom
Loom throws an unclear (and hopefully wrong) error for all the
interleaved tests. It seems to think that the write reference to
the UnsafeCell lasts longer than it does:
```
Causality violation: Concurrent read and write accesses.
    created: /home/kriskras99/Source/ringbeam/src/ring/mod.rs:126:21
       read: thread #1 @ /home/kriskras99/Source/ringbeam/src/ring/recv_values.rs:72:63
```
With [PR loom#279](https://github.com/tokio-rs/loom/pull/279) it also
shows a write in thread #1 but without a location.